Fix the Stripe webhook and amount_raised
The Stripe webhook was broken because it wasn't
the Stripe API key for header verification.

Also, amount_raised was broken because I was
using the wrong seller_id. I was passing in
Seller.seller_id when I should have been
using Seller.id. I added an integration test
to cover this.

Finally I updated the factory for Seller because
using a single "word" was causing a lot of
collisions in the test db.