Distribute the remaining cents between merchants
When calculating how much to pay each merchant,
we can't divide cents. We were previously rounding
up. However, that will cause us to have accounting
errors when trying to do things like "figure out how
much fees we paid Square" or when we want to know
how much money we comped for fees to the merchant.

Therefore we added some logic to distribute the
remaining cents.