Add Fees
We created a new model called "Fee" which allows for
us to apply a fee to the "total amount" it costs per
meal. eg) a $10 meal with a 10% fee would cost $11,
but would generate a $10 gift card. The reasoning for
something like this is to set aside that extra 10% for
things like processing fees or fiscal sponsorships.

This is important for our Think!Chinatown sponsorship
since donations that go through them are tax deductable.