Delegating GAM Payments to Non-Profit Square accounts (Think!Chinatown) (#167)
For Think!Chinatown (and hopefully future partners), we have to delegate payments to their square accounts instead of ours for non-profit type transactions. Right now this is just GAM, but we want to make this generic. The current model as of now is to use one "Send Chinatown Love" Location ID within Think!Chinatown's square account, but we could support other scenarios too.

Overview:

Adding a non_profit_location_id to the sellers model, which represents an alternative square location id to use for non-profit type transactions.
If performing a GAM transaction (is_distribution is true), and the seller has a valid non_profit_location_id set, then use that location id instead of the seller's usual one
Based on that we have to use the right Square credentials for performing the transaction
Supporting details:

Adding more env variables with all the new Think!Chinatown credentials
Making sure we support webhooks from their square account
Appropriate schema and migration for the new field
An new seller to be seeded for testing that supports these scenarios
A more than healthy dose of testing!
Confirmed working payment in Think!Chinatown's sandbox app!

* Added new property "non_profit_location_id" to sellers, and corresponding migration

* Adding check for whether to use non_profit_location_id

* adding check for blank

* fixing test break due to seller schema change

* Added test for non_profit_location_id set for payment_intent

* fixed context blocks for more sensible description

* using Think Chinatown square credentials where appropriate

* running rubocop and adding one more variable to .env.example

* more formatting and test fixes

* update webhook validator test

* test suite config for new env vars

* changing seed location id's to real sandbox location id's

* running rubocop -a

* addressing code review comments

* taking migrations up to date

* adding a unit test for access_token_from_location_id

Co-authored-by: Benjamin Nguyen <benjamin.van.nguyen@gmail.com>