Task to send emails of recent gift cards to merchants (#168)
This is a rake task to email a given seller (and email) vouchers for their businesses over a certain range of time. The default is the last year, which is pretty much "all time" for us. The time is an argument to the task if we want it.

I think this will be short-lived as @justintmckibben is taking this query and building out an awesome endpoint to build a persistent, dedicated front-end site to host this info for merchants to reference. (#175)

As this is probably gonna mostly be a one-off, I've kept it light on the formatting / styling with just a plain email.