Fix broken migration
We removed the relationship to Contact from the Seller model so
when we ran the migration, Seller.distributor indeed does not
exist. We have to get the Contact the old fashioned way without
the model relationship.

Also, Campaign has required fields:
 - seller

So we have to be sure it's there when we created it