Migrate all GAM items to have the right campaign
Previously we denoted a GAM gift card by the fact
that it is single_use. Now that we have added a
Campaign object that runs each campaign, we can use
the presence of that to determine GAM gift cards.
We associated all gift cards of singel use to their
corresponding Campaign on the Seller. Currentlly there
is max one Campaign.