Don't update token when requesting rewards token twice in a row
Sometimes when you double submit to get an email sent,
the endpoint will send multiple emails. This can get
confusing for a user since they won't know which email
has the active token. Therefore we updated it so that
it doesn't create a new token unless the last one is
expired. It will still send the same email though.

Note that each time you hit the endpoint though it will
update the expiration time to give an additional 30 mins
for that token.